SC-134: Manual power off should be independent of protection rules (#52)
* independent power off

* add changeset

+++ include/sensors/classes/acPzem.h
@@ -19,7 +19,7 @@ struct Zone {
 
 class AcPzem {
  public:
-  AcPzem(String name, SoftwareSerial& port, uint8_t storageAddress, uint8_t pzemAddress, uint8_t avgVoltageSize);
+  AcPzem(String name, SoftwareSerial& port, uint8_t storageAddress, uint8_t pzemAddress, uint16_t avgVoltageSize);
 
   void startPzem();
